requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_pytorch_select=0.1=cpu_0
blas=1.0=mkl
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h0d85af4_4
ca-certificates=2021.10.8=h033912b_0
cairo=1.16.0=he43a7df_1008
certifi=2021.10.8=py39h6e9494a_2
cffi=1.15.0=py39hc55c11b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=37.0.1=py39hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
ffmpeg=4.3.2=h4dad6da_0
fontconfig=2.13.94=h10f422b_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
gettext=0.19.8.1=h7937167_1005
giflib=5.2.1=haf1e3a3_0
glib=2.68.4=he49afe7_0
glib-tools=2.68.4=he49afe7_0
gmp=6.2.1=h2e338ed_0
gnutls=3.6.13=h756fd2b_1
graphite2=1.3.13=h2e338ed_1001
harfbuzz=2.9.1=h159f659_1
hdf5=1.10.6=nompi_haae91d6_101
icu=68.2=he49afe7_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2019.4=233
jasper=1.900.1=h636a363_1006
jpeg=9e=hca72f7f_0
kiwisolver=1.3.2=py39he9d5cce_0
lame=3.100=h35c211d_1001
lcms2=2.12=hf1fd2bf_0
libblas=3.9.0=1_h43114d0_netlib
libcblas=3.9.0=3_h7d47d6b_netlib
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=4.0.0=7_5_0_h1a10cd1_23
libgfortran4=7.5.0=h1a10cd1_23
libglib=2.68.4=hd556434_0
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=3_h7d47d6b_netlib
liblapacke=3.9.0=3_h7d47d6b_netlib
libmklml=2019.0.5=0
libopencv=4.5.0=py39_2
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=h87d7836_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
libxml2=2.9.12=h93ec3fd_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
matplotlib=3.5.1=py39hecd8cb5_1
matplotlib-base=3.5.1=py39hfb0c5b7_1
mkl=2019.4=233
mkl-service=2.3.0=py39h9ed2024_0
mkl_fft=1.3.0=py39ha059aab_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_2
nettle=3.6=hedd7734_0
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
numpy=1.20.3=py39h139494b_2
opencv=4.5.0=py39_2
openh264=2.1.1=hfd3ada9_0
openssl=1.1.1o=hfe4f2af_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39h4d6be9b_0
pcre=8.45=he49afe7_0
pillow=9.0.1=py39hde71d04_0
pip=21.2.4=py39hecd8cb5_0
pixman=0.40.0=hbcb3906_0
py-opencv=4.5.0=py39h9cd51e4_2
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.12=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.7.1=cpu_py39h7e2095a_0
pytz=2022.1=pyhd8ed1ab_0
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
scipy=1.5.3=py39h0fe7a32_0
setuptools=61.2.0=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.3=h707629a_0
tk=8.6.11=h3fd3227_1
torchvision=0.8.2=cpu_py39hde629fd_0
tornado=6.1=py39h9ed2024_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39hecd8cb5_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!161.3030=h0d85af4_1
xz=5.2.5=hca72f7f_1
zlib=1.2.12=h4dc903c_2
zstd=1.4.9=h322a384_0
